requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
django-grappelli==3.0.6
django-multiselectfield==0.1.12
django-tinymce==3.6.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
